Move rosetta and protocol related items under one module
