+++ rust/tw_internet_computer/src/protocol/principal.rs
